+++ packages/kiapi-relay/src/kiapi_relay/_types/relay.py
@@ -4,10 +4,13 @@ from typing import Protocol, runtime_checkable
 from .._views.relay_request import RelayRequest
 from .._views.relay_response import RelayResponse
 from .relay_delivery import RelayDelivery
+from .relay_name import RelayName
 
 
 @runtime_checkable
 class Relay(Protocol):
+    name: RelayName
+
     def watch(self) -> AsyncIterator[RelayDelivery]: ...
 
     async def request(