+++ packages/kiapi/src/kiapi/api/__init__.py
@@ -4,6 +4,7 @@ from ._helpers.build_openapi import build_openapi
 from ._helpers.build_train_responses import build_train_responses
 from ._helpers.get_accept import get_accept
 from ._helpers.get_ctx import get_ctx
+from ._helpers.get_relay_runner import get_relay_runner
 from ._helpers.get_worker import get_worker
 from ._helpers.register_capability_endpoints import register_capability_endpoints
 from ._helpers.submit_and_maybe_wait import submit_and_maybe_wait
@@ -18,6 +19,7 @@ __all__ = [
     "build_train_responses",
     "get_accept",
     "get_ctx",
+    "get_relay_runner",
     "get_worker",
     "register_capability_endpoints",
     # ._settings